+++ aquamarine/src/vm_pool.rs
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-use std::sync::Arc;
 use std::task::{Context, Poll};
 
 use futures::{future::BoxFuture, FutureExt};
-use health::{HealthCheck, HealthCheckRegistry};
-use parking_lot::Mutex;
+use health::HealthCheckRegistry;
 
 use peer_metrics::VmPoolMetrics;
 
 use crate::aqua_runtime::AquaRuntime;
+use crate::health::VMPoolHealCheck;
 
 type RuntimeF<RT> = BoxFuture<'static, Result<RT, <RT as AquaRuntime>::Error>>;
 
@@ -188,41 +187,5 @@ impl<RT: AquaRuntime> VmPool<RT> {
         if wake {
             cx.waker().wake_by_ref()
         }
-    }
-}
-
-#[derive(Clone)]
-pub struct VMPoolHealCheck {
-    expected_count: usize,
-    current_count: Arc<Mutex<usize>>,
-}
-
-impl VMPoolHealCheck {
-    pub fn new(expected_count: usize) -> Self {
-        Self {
-            expected_count,
-            current_count: Arc::new(Mutex::new(0)),
-        }
-    }
-
-    pub fn increment_count(&self) {
-        let mut guard = self.current_count.lock();
-        *guard += 1;
-    }
-}
-
-impl HealthCheck for VMPoolHealCheck {
-    fn check(&self) -> eyre::Result<()> {
-        let guard = self.current_count.lock();
-        let current = *guard;
-        if self.expected_count != current {
-            return Err(eyre::eyre!(
-                "VM pool isn't full. Current: {}, Expected: {}",
-                current,
-                self.expected_count
-            ));
-        }
-
-        Ok(())
     }
 }
